+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Explicitly allow reading the `fhirUser` resource when scope and launch context ID are present (RFS-309)
+- Add special handling for `patient/*` scopes for resources outside of the patient compartment (RFS-309)
 
 ## [0.2.0] - 2026-07-13